+++ src/views/home/right/OutputRight.tsx
@@ -37,7 +37,7 @@ const OutputRight: React.FC<Props> = ({ data, dataFull }) => {
       />
       <StaticBlock
         title="Condorcet"
-        methods={['copeland', 'lull', 'kemeny_young']}
+        methods={['copeland', 'lull', 'kemeny_young', 'black']}
         results={dataFull}
       />
       <StaticBlock